Refs #5371. Rename UDP agents.
