+++ include/uxr/agent/utils/CLI.hpp
@@ -16,10 +16,10 @@
 #define UXR_AGENT_UTILS_CLI_HPP_
 
 #ifdef _WIN32
-#include <uxr/agent/transport/udp/UDPServerWindows.hpp>
+#include <uxr/agent/transport/udp/UDPv4AgentWindows.hpp>
 #include <uxr/agent/transport/tcp/TCPServerWindows.hpp>
 #else
-#include <uxr/agent/transport/udp/UDPServerLinux.hpp>
+#include <uxr/agent/transport/udp/UDPv4AgentLinux.hpp>
 #include <uxr/agent/transport/tcp/TCPServerLinux.hpp>
 #include <uxr/agent/transport/serial/SerialServerLinux.hpp>
 #include <uxr/agent/transport/serial/baud_rate_table_linux.h>
